// services/portfolio-service/src/models/regulatory/Regulatory.ts
export enum RegulatoryJurisdiction {
  SEC = 'SEC',
  FINRA = 'FINRA',
  CFTC = 'CFTC',
  STATE = 'STATE',
  INTERNATIONAL = 'INTERNATIONAL'
}